requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
openai>=1.0.0
ddgs>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.30.0
fastapi>=0.110.0
uvicorn>=0.29.0
reportlab>=4.0.0
pydantic>=2.0.0
python-multipart>=0.0.9